Updating to NuGet 4.1 API so that we can handle package name extraction properly
references #20

+++ src/Nanicitus.Core/SymbolIndexer.cs
@@ -11,17 +11,18 @@ using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Globalization;
 using System.IO;
+using System.IO.Compression;
 using System.Linq;
 using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
-using Ionic.Zip;
 using Nanicitus.Core.Monitoring;
 using Nanicitus.Core.Properties;
 using Nuclei.Configuration;
 using Nuclei.Diagnostics;
 using Nuclei.Diagnostics.Logging;
-using NuGet;
+using NuGet.Packaging;
+using NuGet.Packaging.Core;
 
 namespace Nanicitus.Core
 {
@@ -330,6 +331,64 @@ namespace Nanicitus.Core
             }
         }
 
+        /// <summary>
+        /// Extracts the files from a NuGet symbol package.
+        /// </summary>
+        /// <param name="packageFile">The full path to the file that contains the package information.</param>
+        /// <param name="extractDirectory">The directory into which packages can be extracted.</param>
+        /// <param name="onLoadFailure">The action that should be execute if the package fails to load.</param>
+        /// <returns>The directory into which the package was extracted.</returns>
+        [SuppressMessage(
+            "Microsoft.Design",
+            "CA1031:DoNotCatchGeneralExceptionTypes",
+            Justification = "Do not want the application to crash if there is an error loading symbols.")]
+        public string ExtractSymbolPackage(
+            string packageFile,
+            string extractDirectory,
+            Action<string, Exception> onLoadFailure)
+        {
+            try
+            {
+                int waitCount = 0;
+                while ((!PackageUtilities.IsAvailableForReading(packageFile)) && (waitCount < PackageUtilities.MaximumNumberOfTimesWaitingForPackageFileLock))
+                {
+                    waitCount++;
+                    Thread.Sleep(PackageUtilities.PackageFileLockSleepTimeInMilliSeconds);
+                }
+
+                using (var reader = new PackageArchiveReader(new ZipArchive(new FileStream(packageFile, FileMode.Open, FileAccess.Read))))
+                {
+                    var identity = reader.GetIdentity();
+                    var destinationDirectory = Path.Combine(
+                        extractDirectory,
+                        string.Format(
+                            CultureInfo.InvariantCulture,
+                            "{0}.{1}",
+                            identity.Id,
+                            identity.Version.ToNormalizedString()));
+                    if (!Directory.Exists(destinationDirectory))
+                    {
+                        Directory.CreateDirectory(destinationDirectory);
+                    }
+
+                    foreach (var file in reader.GetFiles())
+                    {
+                        reader.ExtractFile(
+                            file,
+                            Path.Combine(destinationDirectory, file),
+                            new NuGetLogger(_diagnostics));
+                    }
+
+                    return destinationDirectory;
+                }
+            }
+            catch (Exception e)
+            {
+                onLoadFailure?.Invoke(packageFile, e);
+                return null;
+            }
+        }
+
         private void HandleOnEnqueue(object sender, EventArgs e)
         {
             CreateWorker();
@@ -444,15 +503,16 @@ namespace Nanicitus.Core
             }
         }
 
-        private bool LoadPackage(out string packageFile, out ZipPackage package)
+#pragma warning disable SA1008 // Opening parenthesis must be spaced correctly
+        private (string packageFile, Action<IndexReport> reportSink, int processCount, PackageIdentity id) LoadPackage()
+#pragma warning restore SA1008 // Opening parenthesis must be spaced correctly
         {
-            packageFile = null;
-            package = null;
-
+            string packageFile = null;
+            Action<IndexReport> reportSink = null;
             int processCount = 0;
             if (!_queue.IsEmpty)
             {
-                packageFile = _queue.Dequeue();
+                (packageFile, reportSink) = _queue.Dequeue();
             }
             else
             {
@@ -467,7 +527,7 @@ namespace Nanicitus.Core
             if (packageFile == null)
             {
                 _diagnostics.Log(LevelToLog.Trace, Resources.Log_Messages_SymbolIndexer_PackageNotDefined);
-                return false;
+                return (packageFile, reportSink, processCount, null);
             }
 
             if (processCount > MaximumNumberOfTimesPackageCanBeProcessed)
@@ -478,23 +538,11 @@ namespace Nanicitus.Core
                         CultureInfo.InvariantCulture,
                         Resources.Log_Messages_PackageLoadCountBeyondMaximum_WithPackagePath,
                         packageFile));
-                return false;
+                return (packageFile, reportSink, processCount, null);
             }
 
-            package = PackageUtilities.LoadSymbolPackage(
-                packageFile,
-                (file, e) =>
-                {
-                    _diagnostics.Log(
-                        LevelToLog.Error,
-                        string.Format(
-                            CultureInfo.InvariantCulture,
-                            Resources.Log_Messages_SymbolIndexer_PackageLoadingFailed_WithException,
-                            e));
-
-                    _lockedPackages.Enqueue(new Tuple<string, int>(file, ++processCount));
-                });
-            return package != null;
+            var id = PackageUtilities.GetPackageIdentity(packageFile);
+            return (packageFile, reportSink, processCount, id);
         }
 
         private void MarkSymbolsAsProcessed(string packageFile, string project, Version version)
@@ -545,15 +593,10 @@ namespace Nanicitus.Core
                         break;
                     }
 
-                    string packageFile;
-                    ZipPackage package;
-                    if (!LoadPackage(out packageFile, out package))
-                    {
-                        continue;
-                    }
+                    (string packageFile, Action<IndexReport> reportSink, int processCount, PackageIdentity id) = LoadPackage();
 
-                    var project = package.Id;
-                    var version = package.Version.Version;
+                    var project = id.Id;
+                    var version = id.Version.Version;
 
                     _diagnostics.Log(
                         LevelToLog.Info,
@@ -565,9 +608,23 @@ namespace Nanicitus.Core
 
                     // Unpack package file in temp location
                     var processingWasSuccessful = true;
-                    var unpackLocation = Unpack(packageFile, project, version);
+                    var unpackLocation = string.Empty;
                     try
                     {
+                        unpackLocation = ExtractSymbolPackage(
+                            packageFile,
+                            _unpackDirectory,
+                            (file, e) =>
+                            {
+                                _diagnostics.Log(
+                                    LevelToLog.Error,
+                                    string.Format(
+                                        CultureInfo.InvariantCulture,
+                                        Resources.Log_Messages_SymbolIndexer_PackageLoadingFailed_WithException,
+                                        e));
+
+                                _lockedPackages.Enqueue(new Tuple<string, int>(file, ++processCount));
+                            });
                         IndexSymbols(unpackLocation, project, version);
                         UploadSources(unpackLocation, project, version);
                         UploadSymbols(unpackLocation, project, version);
@@ -587,13 +644,22 @@ namespace Nanicitus.Core
                                 project,
                                 version));
 
+                        reportSink(
+                            new IndexReport(
+                                project,
+                                version.ToString(),
+                                IndexStatus.Failed,
+                                new[] { e.ToString() }));
                         StoreSymbolProcessMetrics(false);
                     }
                     finally
                     {
                         try
                         {
-                            Directory.Delete(unpackLocation, true);
+                            if (Directory.Exists(unpackLocation))
+                            {
+                                Directory.Delete(unpackLocation, true);
+                            }
                         }
                         catch (IOException e)
                         {
@@ -692,31 +758,6 @@ namespace Nanicitus.Core
                 {
                     { "Success", wasSuccessful.ToString() }
                 });
-        }
-
-        private string Unpack(string packageFile, string project, Version version)
-        {
-            var destinationDirectory = Path.Combine(
-                _unpackDirectory,
-                string.Format(
-                    CultureInfo.InvariantCulture,
-                    "{0}.{1}",
-                    project,
-                    FormatVersion(version)));
-            if (!Directory.Exists(destinationDirectory))
-            {
-                Directory.CreateDirectory(destinationDirectory);
-            }
-
-            using (var zipFile = ZipFile.Read(packageFile))
-            {
-                foreach (var entry in zipFile)
-                {
-                    entry.Extract(destinationDirectory, ExtractExistingFileAction.OverwriteSilently);
-                }
-            }
-
-            return destinationDirectory;
         }
 
         private void UploadSources(string unpackLocation, string project, Version version)
